Move errors in it own module
