+++ osmapi/__init__.py
@@ -3,3 +3,4 @@ from __future__ import (absolute_import, print_function, unicode_literals)
 __version__ = '1.3.0'
 
 from .OsmApi import *  # noqa
+from .errors import *  # noqa
